+++ examples/02_split_api.py
