+++ nnx/__init__.py
@@ -2,7 +2,6 @@ __version__ = "0.0.0"
 
 
 from .context import Context, RngStream
-from .filters import jit_filter_pure
 from .module import Module, ModuleDef, PureModule, PureModule
 from .nn.activations import (
     celu,
