+++ scripts/run-all-examples.bash
@@ -4,5 +4,5 @@ for f in $(find examples -name "*.py"); do
     echo -e "\n---------------------------------"
     echo "$f"
     echo "---------------------------------"
-    poetry run python "$f"
+    poetry run time python "$f"
 done
